requirements.txt
requests==2.32.3
rich==14.0.0
pyyaml==6.0.2